SpotifyFlaskApp/requirements.txt
Flask
spotipy
psycopg2-binary
requests
python-dotenv
gunicorn
colorlog
sqlalchemy